nit: move condig document helper
